dashboard/requirements.txt
uvicorn==0.22.0
fastapi==0.103.2
jinja2==3.1.3
python-multipart==0.0.6
python-dotenv==0.21.1